Fix compatibility with warptorio
After a warp, warptorio deletes old surfaces, which then leads to an
error in check_burner when there were burner inserters in the old world,
because it wants to check if the inserter was replaced. I added an
additional check that also removes burners from the table when the old
surface doesn't exist anymore, because they obviously can't be replaced
then.

But this leads to another problem: Everything on the platform was cloned
before the old surface was removed. So burner inserters on the platform
lost their leech functionality after the clone. So I added another event
that also adds burners to the table when they are cloned.